Add ps2_interrupt.c
