+++ protocol.mk
@@ -13,7 +13,7 @@ ifdef PS2_USE_BUSYWAIT
 endif
 
 ifdef PS2_USE_INT
-    SRC += protocol/ps2.c
+    SRC += protocol/ps2_interrupt.c
     OPT_DEFS += -DPS2_USE_INT
 endif
